+++ lib/expansion_tile_list.dart
@@ -1,4 +1,5 @@
-/// Library for ExpansionTileList and ExpansionTileExtension
+/// ExpansionTileList is a Flutter package that allows you to create a list of ExpansionTile widgets.
+/// It provides a highly customisable list of expansion tiles with extensive options for appearance, animations, and control over expansion behaviour, while simplifying common features.
 library;
 
 export 'src/expansion_tile_extension.dart';